修改门店列表导出以及数据显示、搜索

+++ src/api/Req.js
@@ -6,43 +6,45 @@ const defaultHeader = {
   // 'Accept': 'application/json, text/plain, */*',
   'Content-Type': 'application/json'
 };
-function request({ baseURL = '', timeout = 600000, headers = defaultHeader}) {
+function request({ baseURL = '', timeout = 600000, headers = defaultHeader,isInterceptors = true}) {
   const axiosinstance  = axios.create({
     baseURL,
     timeout,
     headers,
     withCredentials: true,
-  })
-  axiosinstance.interceptors.request.use((config) => {
-    return config;
-  },error => {
-    Promise.reject({
-      message:error.message || '请求参数异常'
-    })
-  })
-  // 请求响应拦截器
-  axiosinstance.interceptors.response.use((response) => {
-      const { httpCode, msg:resultMessage, result} = response.data;
-      // 用户登录超时统一处理
-      if(httpCode=='E_300'){
-         window.location.href= '/';
-         sessionStorage.clear();
-         return;
-      }
-      if(httpCode!=200){//业务错误弹框
-          Qmessage.error(resultMessage);
-          return Promise.reject(result);
-      }
-      return {result,httpCode};
-    }, error => {
-      // window.location.href= '/';
-      sessionStorage.clear();
-      Qmessage.error('服务异常');
-      setTimeout(()=> {
-        // window.location.href= '/';
-      },4000)
-      return Promise.reject({message:'服务异常'});
-    });
+  });
+  if(isInterceptors){
+      axiosinstance.interceptors.request.use((config) => {
+          return config;
+      },error => {
+          Promise.reject({
+              message:error.message || '请求参数异常'
+          })
+      })
+      // 请求响应拦截器
+      axiosinstance.interceptors.response.use((response) => {
+          const { httpCode, msg:resultMessage, result} = response.data;
+          // 用户登录超时统一处理
+          if(httpCode=='E_300'){
+              window.location.href= '/';
+              sessionStorage.clear();
+              return;
+          }
+          if(httpCode!=200){//业务错误弹框
+              Qmessage.error(resultMessage);
+              return Promise.reject(result);
+          }
+          return {result,httpCode};
+      }, error => {
+          // window.location.href= '/';
+          sessionStorage.clear();
+          Qmessage.error('服务异常');
+          setTimeout(()=> {
+              // window.location.href= '/';
+          },4000)
+          return Promise.reject({message:'服务异常'});
+      });
+  }
   return axiosinstance;
 }
 // const ajax = new request({baseURL:'/qtoolsOms'});
@@ -51,8 +53,11 @@ function request({ baseURL = '', timeout = 600000, headers = defaultHeader}) {
 const omsAjax = new request({baseURL:'/qtoolsOms'});
 const erpAjax = new request({baseURL:'/qtoolsErp'});
 const appAjax = new request({baseURL:'/qtoolsApp'});
+const omsEmptyInterceptorsAjax = new request({baseURL:'/qtoolsOms',isInterceptors:false});
+const erpEmptyInterceptorsAjax = new request({baseURL:'/qtoolsErp',isInterceptors:false});
+const appEmptyInterceptorsAjax = new request({baseURL:'/qtoolsApp',isInterceptors:false});
 export {
-  omsAjax,erpAjax,appAjax
+  omsAjax,erpAjax,appAjax,omsEmptyInterceptorsAjax,erpEmptyInterceptorsAjax,appEmptyInterceptorsAjax
 }
 /**
  * 网络请求成功code
